Revert "Replace dotnetzip"
