+++ Demo/DemoItemsForm.cs
@@ -67,6 +67,7 @@ namespace unvell.ReoGrid.Demo
 			var dummyGrid = new ReoGridControl();
 
 			this.labTitle.Text = dummyGrid.ProductName + " " + dummyGrid.ProductVersion.ToString();
+			web.Visible = false;
 
 			// load default demo item
 			if (!string.IsNullOrEmpty(demoFile.defaultItem))
@@ -173,6 +174,28 @@ namespace unvell.ReoGrid.Demo
 			{
 				demoPanel.Controls.Clear();
 				demoPanel.Controls.Add(item._demoInstance);
+
+				if (demoCtrl is IDemoHelp
+					&& !string.IsNullOrEmpty(htmlHelp = ((IDemoHelp)demoCtrl).GetHTMLHelp()))
+				{
+					web.DocumentText = string.Format(
+						unvell.ReoGrid.Demo.Properties.Resources.HTMLHelpTemp, item.name, htmlHelp);
+
+					web.Document.OpenNew(true);
+					web.Visible = true;
+				}
+				else if (item.docUrl != null && !string.IsNullOrEmpty(item.docUrl.val))
+				{
+					web.Navigate(string.Format("{0}/{1}", demoFile.baseSite, item.docUrl.val));
+					web.Visible = true;
+				}
+				else
+				{
+					web.Visible = false;
+					//web.DocumentText = unvell.ReoGrid.Demo.Properties.Resources.NoHelpAvailable;
+					//web.Document.OpenNew(true);
+				}
+			
 			}
 		}
