+++ ReoGrid.sln
@@ -11,7 +11,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DemoJP", "DemoJP\DemoJP.csp
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D0F553D0-23DB-499A-89AF-6264A717D1BC}"
 	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
 		LICENSE = LICENSE
 		README.md = README.md
 	EndProjectSection
