+++ ReoGrid/ReoGrid.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net8.0-windows</TargetFrameworks>
+    <TargetFrameworks>netcoreapp3.1;net472</TargetFrameworks>
     <OutputType>Library</OutputType>
     <RootNamespace>unvell.ReoGrid</RootNamespace>
     <AssemblyName>unvell.ReoGrid</AssemblyName>
@@ -156,6 +156,9 @@
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="Resources\EULA_EN.txt" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="DotNetZip" Version="1.15.0" />
   </ItemGroup>
   <ItemGroup>
     <None Update="Properties\Settings.settings">
